deepCCS/DeepCollisionalCrossSection-train/requirements.txt

Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
natsort==6.0.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numba==0.46.0
numexpr==2.7.0
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.2
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.1.1
pyzmq==18.1.0
qtconsole==4.5.5
scanpy==1.4.4.post1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
shap==0.30.1
six==1.12.0
sklearn==0.0
statsmodels==0.10.1
tables==3.5.2
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
umap-learn==0.3.10
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0